fix Pic model, add hash check in file-remove function

+++ src/db/models/Static2.js
@@ -64,14 +64,19 @@ module.exports = function (rootpath)
 
 	function remove_file (hash)
 	{
-		return exists(hash)
+		if (! hash)
+		{
+			return
+		}
+
+		var t = tuple(hash)
+		var path = tuple_to_filename(t)
+
+		return exists(path)
 		.then(exists =>
 		{
 			if (exists)
 			{
-				var t = tuple(hash)
-				var path = tuple_to_filename(t)
-
 				return unlink(path)
 			}
 		})
